+++ volume/api/delete.go
@@ -7,7 +7,7 @@ import (
 	"github.com/syndtr/goleveldb/leveldb/errors"
 )
 
-func Delete(host string, port int, vid int, fid uint64, filename string) error {
+func Delete(host string, port int, vid uint64, fid uint64, filename string) error {
 	url := fmt.Sprintf("http://%s:%d/%d/%d/%s", host, port, vid, fid, filename)
 	req, _ := http.NewRequest(http.MethodDelete, url, nil)
